refactor: :recycle: rename verify function files
