+++ sprout/core/verify_package_properties.py
@@ -1,7 +1,7 @@
 from frictionless.errors import PackageError
 
-from sprout.core.verify_properties_complete import verify_properties_complete
-from sprout.core.verify_properties_well_formed import verify_properties_well_formed
+from sprout.core.verify_properties_are_complete import verify_properties_complete
+from sprout.core.verify_properties_are_well_formed import verify_properties_well_formed
 
 REQUIRED_PACKAGE_PROPERTIES = {
     "name",
