+++ sprout/core/verify_properties_are_complete.py
