+++ sprout/core/verify_properties_are_well_formed.py
