+++ tests/core/test_verify_properties_are_complete.py
@@ -2,7 +2,7 @@ from frictionless.errors import PackageError
 from pytest import mark, raises
 
 from sprout.core.not_properties_error import NotPropertiesError
-from sprout.core.verify_properties_are_complete import verify_properties_are_complete
+from sprout.core.verify_properties_are_complete import verify_properties_complete
 
 
 @mark.parametrize("required_fields", [{}, {"field1"}, {"field1", "field2"}])
@@ -16,7 +16,7 @@ def test_accepts_required_fields(required_fields):
     }
 
     assert (
-        verify_properties_are_complete(properties, PackageError, required_fields)
+        verify_properties_complete(properties, PackageError, required_fields)
         == properties
     )
 
@@ -26,7 +26,7 @@ def test_rejects_empty_object():
     required_fields = {"field1", "field2"}
 
     with raises(NotPropertiesError) as error:
-        verify_properties_are_complete({}, PackageError, required_fields)
+        verify_properties_complete({}, PackageError, required_fields)
 
     message = str(error.value)
     for field in required_fields:
@@ -41,4 +41,4 @@ def test_rejects_empty_value_for_required_fields(empty_value):
     }
 
     with raises(NotPropertiesError, match="'field1' is a required property"):
-        verify_properties_are_complete(properties, PackageError, {"field1"})
+        verify_properties_complete(properties, PackageError, {"field1"})
