+++ tests/core/test_verify_properties_are_well_formed.py
@@ -3,7 +3,7 @@ from pytest import fixture, raises
 
 from sprout.core.not_properties_error import NotPropertiesError
 from sprout.core.properties import PackageProperties, ResourceProperties
-from sprout.core.verify_properties_are_well_formed import verify_properties_are_well_formed
+from sprout.core.verify_properties_are_well_formed import verify_properties_well_formed
 
 error_type = errors.PackageError.type
 
@@ -24,13 +24,13 @@ def test_accepts_default_values():
     """Should accept an object with default values, some of which are blank."""
     properties = PackageProperties().asdict
 
-    assert verify_properties_are_well_formed(properties, error_type) == properties
+    assert verify_properties_well_formed(properties, error_type) == properties
 
 
 def test_accepts_custom_values(package_properties):
     """Should accept a well-formed properties object."""
     assert (
-        verify_properties_are_well_formed(package_properties, error_type)
+        verify_properties_well_formed(package_properties, error_type)
         == package_properties
     )
 
@@ -40,7 +40,7 @@ def test_rejects_properties_not_conforming_to_spec(package_properties):
     package_properties["name"] = "an invalid name with spaces"
 
     with raises(NotPropertiesError, match="at property 'name'"):
-        verify_properties_are_well_formed(package_properties, error_type)
+        verify_properties_well_formed(package_properties, error_type)
 
 
 def test_filters_for_the_specified_error_type(package_properties):
@@ -49,6 +49,6 @@ def test_filters_for_the_specified_error_type(package_properties):
     package_properties["resources"].append(bad_resource)
 
     assert (
-        verify_properties_are_well_formed(package_properties, error_type)
+        verify_properties_well_formed(package_properties, error_type)
         == package_properties
     )
